Move the salesforce upload code into salesforce and return a url.
